Added routing to index
